# openapi.yaml
...
openapi: 3.0.3
info:
  title: ScholarAI API
  description: AI-powered scholarship matching platform
  version: 5.0.0
  contact:
    name: ScholarAI Team
    email: support@scholarai.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.scholarai.com
    description: Production server
paths:
...
  /deduplication/deduplicate:
    post:
      tags:
        - Deduplication
      summary: Deduplicate scholarship items
      description: Remove duplicate scholarship items using hybrid AI and rule-based approach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScholarshipItem'
                  description: Array of scholarship items to deduplicate
                method:
                  type: string
                  enum: [ai, rules, hybrid]
                  default: hybrid
                  description: Deduplication method to use
              required:
                - items
      responses:
        '200':
          description: Deduplication completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeduplicationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/deduplicate-ai:
    post:
      tags:
        - Deduplication
      summary: Deduplicate using AI only
      description: Remove duplicate scholarship items using AI-powered analysis only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScholarshipItem'
                  description: Array of scholarship items to deduplicate
              required:
                - items
      responses:
        '200':
          description: AI deduplication completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeduplicationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/deduplicate-rules:
    post:
      tags:
        - Deduplication
      summary: Deduplicate using rules only
      description: Remove duplicate scholarship items using rule-based similarity matching only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScholarshipItem'
                  description: Array of scholarship items to deduplicate
              required:
                - items
      responses:
        '200':
          description: Rule-based deduplication completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeduplicationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/batch-deduplicate:
    post:
      tags:
        - Deduplication
      summary: Batch deduplicate multiple arrays
      description: Deduplicate multiple arrays of scholarship items in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batches:
                  type: array
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScholarshipItem'
                  description: Array of arrays of scholarship items to deduplicate
                method:
                  type: string
                  enum: [ai, rules, hybrid]
                  default: hybrid
                  description: Deduplication method to use
              required:
                - batches
      responses:
        '200':
          description: Batch deduplication completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeduplicationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/validate:
    post:
      tags:
        - Deduplication
      summary: Validate deduplication results
      description: Validate the results of a deduplication operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                original:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScholarshipItem'
                  description: Original array of scholarship items
                deduplicated:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScholarshipItem'
                  description: Deduplicated array of scholarship items
              required:
                - original
                - deduplicated
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/test:
    get:
      tags:
        - Deduplication
      summary: Test deduplication system
      description: Test the deduplication system with sample data
      parameters:
        - in: query
          name: method
          schema:
            type: string
            enum: [ai, rules, hybrid]
            default: hybrid
          description: Deduplication method to test
      responses:
        '200':
          description: Test completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  testData:
                    type: object
                    properties:
                      input:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScholarshipItem'
                      output:
                        $ref: '#/components/schemas/DeduplicationResult'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deduplication/schema:
    get:
      tags:
        - Deduplication
      summary: Get deduplication schema
      description: Retrieve the API schema for deduplication endpoints
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  schema:
                    type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Additional error details (e.g., validation errors)
...
    ScholarshipItem:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the scholarship item
        name:
          type: string
          description: Name of the scholarship
        country:
          type: string
          nullable: true
          description: Country where the scholarship is available
        degree:
          type: string
          enum: ["Bachelor", "Master", "PhD", "Any"]
          description: Degree level for the scholarship
        eligibility:
          type: string
          description: Eligibility requirements
        deadline:
          type: string
          nullable: true
          description: Application deadline
        link:
          type: string
          format: uri
          description: Official application or information URL
        source:
          type: string
          description: Source domain of the scholarship information
      required:
        - id
        - name
        - degree
        - eligibility
        - link
        - source
    DeduplicationResult:
      type: object
      properties:
        success:
          type: boolean
          description: Whether deduplication was successful
        originalCount:
          type: integer
          description: Number of items in original array
        deduplicatedCount:
          type: integer
          description: Number of items after deduplication
        duplicatesRemoved:
          type: integer
          description: Number of duplicate items removed
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScholarshipItem'
          description: Deduplicated array of scholarship items
        duplicateGroups:
          type: array
          items:
            type: object
            properties:
              representative:
                $ref: '#/components/schemas/ScholarshipItem'
              duplicates:
                type: array
                items:
                  $ref: '#/components/schemas/ScholarshipItem'
              similarity:
                type: number
                minimum: 0
                maximum: 1
          description: Groups of duplicates found
        method:
          type: string
          enum: [ai, rules, hybrid]
          description: Method used for deduplication
        metadata:
          type: object
          properties:
            processingTimeMs:
              type: integer
            timestamp:
              type: string
              format: date-time
            model:
              type: string
              description: AI model used (if applicable)
        error:
          type: string
          nullable: true
          description: Error message if deduplication failed
      required:
        - success
        - originalCount
        - deduplicatedCount
        - duplicatesRemoved
        - items
        - method
        - metadata
    BatchDeduplicationResult:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeduplicationResult'
        summary:
          type: object
          properties:
            totalBatches:
              type: integer
            totalOriginalItems:
              type: integer
            totalDeduplicatedItems:
              type: integer
            totalDuplicatesRemoved:
              type: integer
            processingTimeMs:
              type: integer
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - results
        - summary
        - timestamp
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
          description: Whether the deduplication results are valid
        originalCount:
          type: integer
          description: Count of original items
        deduplicatedCount:
          type: integer
          description: Count of deduplicated items
        errors:
          type: array
          items:
            type: string
          description: List of validation errors
        warnings:
          type: array
          items:
            type: string
          description: List of validation warnings
        metadata:
          type: object
          properties:
            duplicatesExpected:
              type: integer
            duplicatesFound:
              type: integer
            integrityCheck:
              type: boolean
        timestamp:
          type: string
          format: date-time
      required:
        - valid
        - originalCount
        - deduplicatedCount
        - errors
        - warnings
        - metadata
        - timestamp
...